refactor: pos-gmp.flows for supplyToAave, supplyToCompound, etc.
